Refactor ReplicationServer handlers
